Fix an unexpected error thrown by TrackerGroup#finish
`TrackerGroup#finish` attempts to call the `finish` method on all of
its children. This worked for Tracker and TrackerGroup instances, but
failed on TrackerStream objects as they do not expose a `finish`
method.

This changeset exposes a `finish` method on TrackerStream instances by
adding "finish" to the method delegation list so any calls will
properly delegate to the underlying tracker instance.